Change ManuService to MenuCLI and move it to new dictionary UI
